updated `aws` autocomplete for DynamicPreview
