+++ shell/tab.go
@@ -1,7 +1,6 @@
 package shell
 
 import (
-	"context"
 	"regexp"
 	"strconv"
 	"strings"
@@ -308,39 +307,4 @@ func cacheHints(prefix string, exes []string) []string {
 	}
 
 	return hints
-}
-
-func DynamicPreview(previewBlock string, exe string, params []string) readline.PreviewFuncT {
-	block := []rune(previewBlock)
-	return func(ctx context.Context, line []rune, item string, _ bool, size *readline.PreviewSizeT, callback readline.PreviewFuncCallbackT) {
-		fork := lang.ShellProcess.Fork(lang.F_FUNCTION | lang.F_BACKGROUND | lang.F_NO_STDIN | lang.F_CREATE_STDOUT | lang.F_NO_STDERR)
-		fork.Name.Set(exe)
-		fork.Parameters.DefineParsed(params)
-		fork.FileRef = autocomplete.ExesFlagsFileRef[exe]
-
-		_, err := fork.Execute(block)
-		if err != nil {
-			s, _, err := previewError(err, size)
-			callback(s, 0, err)
-			return
-		}
-
-		b, err := fork.Stdout.ReadAll()
-		if err != nil {
-			s, _, err := previewError(err, size)
-			callback(s, 0, err)
-			return
-		}
-
-		s, _, err := previewParse(b, size)
-		if err != nil {
-			s, _, err = previewError(err, size)
-			callback(s, 0, err)
-			return
-		}
-
-		i := previewPos(s, item)
-
-		callback(s, i, err)
-	}
 }
